修改RecorderManager函数
